create a dummy alembic_version class
In the tests I was using db.drop_all() to delete the tables and create a
fresh db. This was a problem as the alembic_version table would not be
deleted. Two options: use a specific test db (probably a good idea) or
create a dummy alembic_version class, that way the DB sees the table
link and attempts to delete it when calling drop_all.